arm/configs: Temporarily add Axon-specific defconfig
Add an Axon-specific defconfig for development purposes only. This is to
better isolate and identify issues with wireless configuration by
removing certain kernel configurations such as BRCM-related kernel modules,
and the FW_LOADER_USER_HELPER. Ideally it is best to have a kernel that
supports most wireless options and multiple SOM families to minimize
maintenance. It is anticipated that any validatated changes to this config
will be incorporated into a single configuration in the future.