axon:device-tree: Separately control GPIO pinctrl
To be able to separately control the GPIO on a baseboard-level, we separately
define a pinctrl group for each GPIO. This way the baseboard dts/dtsi determines
which GPIO pins are assigned to pinctrl-0 of the iomux.

Additionally, add a definition for the default pinctrl groups that will be
assigned to pinctrl-0 on the SOM (AXON_SOM_PINCTRL) so that pinctrl-0 can be
overridden in the baseboard dts/dtsi by including it as the first member of
the list. E.g:

&iomuxc {
        pinctrl-0 = AXON_SOM_PINCTRL,
            <&pinctrl_gpio_x1_p66>,
            <&pinctrl_gpio_x1_p68>;
};